backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.26.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
email-validator>=2.0.0

